refactor: split into components
